version 0.8: working PCB; not sure whether UEXT pinout is correct/useful
I want to use this with a RTF5350F-OLinuXino-UM; but on the evaluation
board's UEXT port, the serial RX and TX lines are the same ones used
for the Linux console in the OpenWRT firmware (also physically wired
to the 3-pin UART0 port).  It's thus not a good choice for connecting
peripheral devices, because I don't want to disable the serial console
and risk being unable to access the Linux system, in case anything
goes wrong with network configuration.  But there's another serial port
on the 40-pin EXT4 connector.  That's a better one to use.  So maybe
the next version of this board could have both connectors so that if
one populates a 40-pin female header connector, it can plug right in.

There is no UEXT extension capability to connect anything else at the
same time: the serial port has to be dedicated anyway, because it seems
that having a chip-select pin is not possible (unless I can repurpose
something else that is not being used, as a GPIO).  On the 40-pin
connector it would be possible though.

There is no isolation: DC ground gets connected to neutral, and
line gets connected via a resistor divider to measure voltage.
In theory that should be OK on a 120VAC US-type power line, as long as
the DC power supply is OK with it, because ethernet is transformer-coupled:
https://electronics.stackexchange.com/questions/27756/why-are-ethernet-rj45-sockets-magnetically-coupled
But it could be problematic in case of a 230VAC system where the
voltage and current we want to measure are between two "phase" lines:
then the ground plane of the circuit will be "hot" relative
to the neutral line, which won't be connected.  A different design
is needed.

Thanks to whatnick and https://github.com/whatnick/CS5490_Breakout
for providing the CS5490 schematic symbol for KiCad.  This started
as a fork of that, but now of course it's quite different so I don't
call it a fork anymore.